+++ activerecord/lib/active_record/persistence.rb
@@ -1217,7 +1217,7 @@ module ActiveRecord
       )
     end
 
-    def create_or_update(**nparams, &block)
+    def create_or_update(**, &block)
       _raise_readonly_record_error if readonly?
       return false if destroyed?
       result = new_record? ? _create_record(&block) : _update_record(&block)
