+++ activerecord/test/models/building.rb
@@ -1,0 +1,4 @@
+class Building < ActiveRecord::Base
+  has_many :sales, autosave: true
+  has_many :expenses, autosave: true
+end
