+++ activerecord/test/models/child.rb
@@ -1,0 +1,3 @@
+class Child < ActiveRecord::Base
+  belongs_to :parent, inverse_of: :child, autosave: true
+end
