+++ activerecord/test/models/expense.rb
@@ -1,0 +1,3 @@
+class Expense < ActiveRecord::Base
+  belongs_to :building, autosave: true
+end
