+++ activerecord/test/models/grandparent.rb
@@ -1,0 +1,3 @@
+class Grandparent < ActiveRecord::Base
+  has_one :parent, inverse_of: :grandparent, autosave: true
+end
