+++ activerecord/test/models/listing.rb
@@ -1,0 +1,3 @@
+class Listing < ActiveRecord::Base
+  belongs_to :building, autosave: true
+end
