+++ activerecord/test/models/sale.rb
@@ -1,0 +1,3 @@
+class Sale < ActiveRecord::Base
+  belongs_to :building, autosave: true
+end
